backend/app/services/medgemma_service.py
import torch
from PIL import Image
import numpy as np
from typing import Dict, List, Optional, Union
from loguru import logger
from app.core.config import settings

class MedGemmaService:
    """Service for handling MedGemma AI analysis"""
    
    def __init__(self):
        self.device = torch.device(settings.MEDGEMMA_DEVICE)
        self.model = self._load_model()
        self.confidence_threshold = settings.MEDGEMMA_CONFIDENCE_THRESHOLD
        
    def _load_model(self):
        """Load MedGemma model"""
        try:
            # TODO: Implement actual model loading
            # This is a placeholder for the actual implementation
            logger.info("Loading MedGemma model...")
            return None
        except Exception as e:
            logger.error(f"Failed to load MedGemma model: {str(e)}")
            raise
    
    async def analyze_medical_image(
        self,
        image: Image.Image,
        context: Optional[Dict] = None
    ) -> Dict:
        """Analyze medical image using MedGemma"""
        try:
            # TODO: Implement actual image analysis
            # This is a placeholder for the actual implementation
            return {
                "findings": [],
                "confidence": 0.0,
                "recommendations": []
            }
        except Exception as e:
            logger.error(f"Image analysis failed: {str(e)}")
            raise
    
    async def analyze_clinical_text(
        self,
        text: str,
        context: Optional[Dict] = None
    ) -> Dict:
        """Analyze clinical text using MedGemma"""
        try:
            # TODO: Implement actual text analysis
            # This is a placeholder for the actual implementation
            return {
                "analysis": {},
                "confidence": 0.0,
                "recommendations": []
            }
        except Exception as e:
            logger.error(f"Text analysis failed: {str(e)}")
            raise
    
    async def comprehensive_medical_analysis(
        self,
        image: Optional[Image.Image] = None,
        text: Optional[str] = None,
        vital_signs: Optional[Dict] = None,
        patient_history: Optional[Dict] = None
    ) -> Dict:
        """Perform comprehensive medical analysis using all available data"""
        try:
            results = {
                "image_analysis": {},
                "text_analysis": {},
                "vital_signs_analysis": {},
                "risk_assessment": {},
                "emergency_flags": [],
                "recommendations": []
            }
            
            # Analyze image if provided
            if image:
                results["image_analysis"] = await self.analyze_medical_image(
                    image,
                    context={"patient_history": patient_history}
                )
            
            # Analyze text if provided
            if text:
                results["text_analysis"] = await self.analyze_clinical_text(
                    text,
                    context={"patient_history": patient_history}
                )
            
            # Analyze vital signs if provided
            if vital_signs:
                results["vital_signs_analysis"] = self._analyze_vital_signs(
                    vital_signs,
                    context={"patient_history": patient_history}
                )
            
            # Generate risk assessment
            results["risk_assessment"] = self._generate_risk_assessment(
                results,
                patient_history
            )
            
            # Check for emergency conditions
            results["emergency_flags"] = self._check_emergency_conditions(results)
            
            # Generate recommendations
            results["recommendations"] = self._generate_recommendations(results)
            
            return results
            
        except Exception as e:
            logger.error(f"Comprehensive analysis failed: {str(e)}")
            raise
    
    def _analyze_vital_signs(
        self,
        vital_signs: Dict,
        context: Optional[Dict] = None
    ) -> Dict:
        """Analyze vital signs"""
        # TODO: Implement vital signs analysis
        return {}
    
    def _generate_risk_assessment(
        self,
        analysis_results: Dict,
        patient_history: Optional[Dict] = None
    ) -> Dict:
        """Generate risk assessment based on all available data"""
        # TODO: Implement risk assessment
        return {}
    
    def _check_emergency_conditions(self, analysis_results: Dict) -> List[str]:
        """Check for emergency conditions"""
        # TODO: Implement emergency condition checking
        return []
    
    def _generate_recommendations(self, analysis_results: Dict) -> List[Dict]:
        """Generate recommendations based on analysis results"""
        # TODO: Implement recommendation generation
        return [] 