backend/requirements/base.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# AI and ML
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2

# Medical Image Processing
pydicom==2.4.3
opencv-python==4.8.1.78

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
loguru==0.7.2 